refactor(noon-enabled): rewrite the folders to follow the models
